# endpoints.js
module.exports = function (app) {

    app.get('/users/:id', (req, res) => {

        const filtro = req.query.filtro

        return res.status(404).send(false)

    })

    app.post('/users/:id', (req, res) => {

        const filtro = req.query.filtro

        return res.status(404).send(false)

    })

    app.put('/users/:id', (req, res) => {

        const filtro = req.query.filtro

        return res.status(404).send(false)

    })

   
}